# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a Loan Application.",
        "title": "Loan App API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Ghalmas Shanditya Putra Agung",
            "email": "ghalmas.shanditya.putra.agung@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "paths": {
        "/auth/sign-in": {
            "post": {
                "description": "Authenticate user with username and password to get access and refresh tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign in user",
                "parameters": [
                    {
                        "description": "Sign in credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SignInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SignInResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.SignInRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "Password for authentication (must be strong password)\nrequired: true\nexample: \"StrongP@ssw0rd123\"",
                    "type": "string"
                },
                "username": {
                    "description": "Username for authentication (4-100 characters)\nrequired: true\nmin: 4\nmax: 100\nexample: \"john.doe\"",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 4
                }
            }
        },
        "model.SignInResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "JWT access token for API authentication\nexample: \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c\"",
                    "type": "string"
                },
                "refresh_token": {
                    "description": "JWT refresh token for obtaining new access tokens\nexample: \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c\"",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "bearer": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}